Move preferences in a separate main level folder
