+++ test/functional-api/preferences/hub-preferences.it-spec.ts
@@ -17,8 +17,6 @@ import {
   removeUserAsHubAdmin,
   userAsHubAdminVariablesData,
 } from '@test/utils/mutations/authorization-mutation';
-import { createOrgAndHubWithUsers } from '../create-entities-with-users-helper';
-import { entitiesId, users } from '../communications-helper';
 import {
   createTestHub,
   getHubData,
@@ -26,6 +24,8 @@ import {
 } from '@test/functional-api/integration/hub/hub.request.params';
 import { deleteOrganization } from '@test/functional-api/integration/organization/organization.request.params';
 import { joinCommunity } from '@test/functional-api/user-management/application/application.request.params';
+import { createOrgAndHubWithUsers } from '../zcommunications/create-entities-with-users-helper';
+import { entitiesId, users } from '../zcommunications/communications-helper';
 
 let organizationName = 'h-pref-org-name' + uniqueId;
 let hostNameId = 'h-pref-org-nameid' + uniqueId;
